// src/modules/enum/dashboard-capability.enum.ts
export enum EDashboardCapability {
  EDIT_PROFILE = 'EDIT_PROFILE',
  EDIT_PASSWORD = 'EDIT_PASSWORD',
  // IMAGE
  CREATE_IMAGE = 'CREATE_IMAGE',
  READ_IMAGE = 'READ_IMAGE',
  DELETE_IMAGE = 'DELETE_IMAGE',
  // MENU_CATEGORY
  CREATE_MENU_CATEGORY = 'CREATE_MENU_CATEGORY',
  READ_MENU_CATEGORY = 'READ_MENU_CATEGORY',
  UPDATE_MENU_CATEGORY = 'UPDATE_MENU_CATEGORY',
  DELETE_MENU_CATEGORY = 'DELETE_MENU_CATEGORY',
  // MENU_PRODUCT
  CREATE_MENU_PRODUCT = 'CREATE_MENU_PRODUCT',
  READ_MENU_PRODUCT = 'READ_MENU_PRODUCT',
  UPDATE_MENU_PRODUCT = 'UPDATE_MENU_PRODUCT',
  DELETE_MENU_PRODUCT = 'DELETE_MENU_PRODUCT',
  // ORDER_LOCATION
  CREATE_ORDER_LOCATION = 'CREATE_ORDER_LOCATION',
  READ_ORDER_LOCATION = 'READ_ORDER_LOCATION',
  UPDATE_ORDER_LOCATION = 'UPDATE_ORDER_LOCATION',
  DELETE_ORDER_LOCATION = 'DELETE_ORDER_LOCATION',
  // ORDER_LOCATION
  CREATE_ACCOUNTING_ENTER = 'CREATE_ACCOUNTING_ENTER',
  READ_ACCOUNTING_ENTER = 'READ_ACCOUNTING_ENTER',
  // ACCOUNTING_EXIT
  CREATE_ACCOUNTING_EXIT = 'CREATE_ACCOUNTING_EXIT',
  READ_ACCOUNTING_EXIT = 'READ_ACCOUNTING_EXIT',
  UPDATE_ACCOUNTING_EXIT = 'UPDATE_ACCOUNTING_EXIT',
  DELETE_ACCOUNTING_EXIT = 'DELETE_ACCOUNTING_EXIT',
  // ACCOUNTING_ENTER
  CREATE_WAREHOUSE_ENTER = 'CREATE_WAREHOUSE_ENTER',
  READ_WAREHOUSE_ENTER = 'READ_WAREHOUSE_ENTER',
  UPDATE_WAREHOUSE_ENTER = 'UPDATE_WAREHOUSE_ENTER',
  DELETE_WAREHOUSE_ENTER = 'DELETE_WAREHOUSE_ENTER',
  // WAREHOUSE_EXIT
  CREATE_WAREHOUSE_EXIT = 'CREATE_WAREHOUSE_EXIT',
  READ_WAREHOUSE_EXIT = 'READ_WAREHOUSE_EXIT',
  UPDATE_WAREHOUSE_EXIT = 'UPDATE_WAREHOUSE_EXIT',
  DELETE_WAREHOUSE_EXIT = 'DELETE_WAREHOUSE_EXIT',
  // ECONOMIC_PACKAGE
  CREATE_ECONOMIC_PACKAGE = 'CREATE_ECONOMIC_PACKAGE',
  READ_ECONOMIC_PACKAGE = 'READ_ECONOMIC_PACKAGE',
  DELETE_ECONOMIC_PACKAGE = 'READ_ECONOMIC_PACKAGE',
  UPDATE_ECONOMIC_PACKAGE = 'READ_ECONOMIC_PACKAGE',
  // BLOG
  CREATE_BLOG = 'CREATE_BLOG',
  READ_BLOG = 'READ_BLOG',
  UPDATE_BLOG = 'UPDATE_BLOG',
  DELETE_BLOG = 'DELETE_BLOG',
  // ORDER
  CREATE_ORDER = 'CREATE_ORDER',
  READ_ORDER = 'READ_ORDER',
  UPDATE_ORDER = 'UPDATE_ORDER',
  DELETE_ORDER = 'DELETE_ORDER',
  // FACTOR
  CREATE_FACTOR = 'CREATE_FACTOR',
  READ_FACTOR = 'READ_FACTOR',
  UPDATE_FACTOR = 'UPDATE_FACTOR',
  DELETE_FACTOR = 'DELETE_FACTOR',
  // ONLINE SHOP
  READ_ALL_ONLINE_SHOP = 'READ_ALL_ONLINE_SHOP',
  READ_SNAP_ONLINE_SHOP = 'READ_SNAP_ONLINE_SHOP',
  READ_TAPSI_ONLINE_SHOP = 'READ_TAPSI_ONLINE_SHOP',
  // SUPPORT
  READ_USER = 'READ_USER',
  DELETE_USER = 'DELETE_USER',
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  READ_ROLE = 'READ_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',
  CREATE_ROLE = 'CREATE_ROLE',
  UPDATE_ROLE = 'UPDATE_ROLE',
  // SERVICE
  SEND_SMS_WELCOME = 'SEND_SMS_WELCOME',
}