// src/modules/enum/message-http-exception.enum.ts
export enum EMessageHttpException {
  TOKEN_NOT_FOUND = 'توکن وجود ندارد',
  TOKEN_IS_EXPIRES = 'توکن منقضی شده',
  TOKEN_IS_NOT_EXPIRES = 'توکن منقضی نشده',
  LOGIN_AGAIN = 'ورود مجدد',
  USER_NOT_FOUND = 'کاربر وجود ندارد',
  OTP_FILED = 'کد یکبار مصرف تایید نشد',
  USER_FOUND = 'کاربر وجود دارد',
  ACCESS_DENIED = 'دسترسی وجود ندارد',
  PASSWORD_INCORRECT = 'رمز ورود اشتباه هست',
  IMAGE_NOT_UPLOADED = 'عکس آپلود نشد',
  IMAGE_NOT_DELETE = 'عکس حذف نشد',
  FILED_NOT_FOUND = "اطلاعات کامل نیست"
}